Add files via upload
This project is part of my internship with Oasis Infobyte. Out of the three tasks chosen, this calculator is the first task.

For this poject, I have chosen the glassmorphism UI and CSS grid. The calculation part is done using the eval() function of JS. Taken from microsoft calculator as an inspiration, I have also tried to recreate the history section of past calculations. Also used JQuery to append the li items for the history section.